code

+++ routes/categories.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const categoriesController = require('../controllers/categories');
+const { verifyToken, check_authorization } = require('../utils/check_auth');
 
 router.get('/', async (req, res) => {
   const categories = await categoriesController.getAllCategories();
@@ -12,17 +13,23 @@ router.get('/:id', async (req, res) => {
   res.json(category);
 });
 
-router.post('/', async (req, res) => {
+router.post('/',verifyToken,
+  check_authorization(['admin']),
+  async (req, res) => {
   const created = await categoriesController.createCategory(req.body);
   res.json(created);
 });
 
-router.put('/:id', async (req, res) => {
+router.put('/:id',verifyToken,
+  check_authorization(['admin']),
+  async (req, res) =>{
   const updated = await categoriesController.updateCategory(req.params.id, req.body);
   res.json(updated);
 });
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id',verifyToken,
+  check_authorization(['admin']),
+  async (req, res) => {
   const deleted = await categoriesController.deleteCategory(req.params.id);
   res.json(deleted);
 });
